chore: move usePagination hook outside GenericTable folder (#39997)
