+++ packages/ui-client/src/components/GenericTable/hooks/index.ts
@@ -1,6 +1,6 @@
-export * from './useCurrent';
-export * from './useItemsPerPage';
-export * from './useItemsPerPageLabel';
-export * from './usePagination';
-export * from './useShowingResultsLabel';
+export * from '../../../hooks/useCurrent';
+export * from '../../../hooks/useItemsPerPage';
+export * from '../../../hooks/useItemsPerPageLabel';
+export * from '../../../hooks/usePagination';
+export * from '../../../hooks/useShowingResultsLabel';
 export * from './useSort';
