+++ packages/ui-client/src/hooks/index.ts
@@ -1,3 +1,4 @@
+export * from './useCurrent';
 export * from './useDefaultSettingFeaturePreviewList';
 export * from './useDocumentTitle';
 export * from './useDontAskAgain';
@@ -5,8 +6,13 @@ export * from './useEmbeddedLayout';
 export * from './useFeaturePreview';
 export * from './useFeaturePreviewList';
 export * from './useGoToDirectMessage';
+export * from './useItemsPerPage';
+export * from './useItemsPerPageLabel';
 export * from './useLicense';
+export * from './usePagination';
 export * from './usePreferenceFeaturePreviewList';
+export * from './useShowingResultsLabel';
 export * from './useThemeMode';
 export * from './useUserDisplayName';
 export * from './useValidatePassword';
+
