+++ packages/ui-client/src/hooks/usePagination.ts
@@ -5,9 +5,6 @@ import { useItemsPerPage } from './useItemsPerPage';
 import { useItemsPerPageLabel } from './useItemsPerPageLabel';
 import { useShowingResultsLabel } from './useShowingResultsLabel';
 
-/**
- * TODO: Move `usePagination` outside from `GenericTable` folder
- */
 export const usePagination = (): {
 	current: ReturnType<typeof useCurrent>[0];
 	setCurrent: ReturnType<typeof useCurrent>[1];
